Lots of refactoring on the question and renderer class.
This is much closer to being a working Moodle question type, although
there are still more issues to resolve.

This passes what unit tests we have, which just goes to show we need
more tests! That is the next stage. Add more stringent tests to prove
that everything is working, and fix the code to make them pass.