Interaction elements now live in separate folders.
This makes it easier to add new ones. The can be developed as separate
git projects, and cloned into stack in the same way that we clone the
stack qtype into a Moodle install.

There is no longer a hard-coded list of konwn input types. The ones
installed are deteced automatically.