Inputs are now validated in two phases if necessary.
Annoyingly, Moodle's adaptive mode displays a message "The submission
was invalid, and has been disregarded without penalty." after the first
submit, which is ugly. We should be able to get that re-worded in Moodle
core.

As part of this, stack_input::get_xhtml has been reworded to render, and
the time at which we generate the display of the validation has been
changed from in validate_student_response to in a separate
render_validation method that is called from the question type renderer.

I also cleaned up most of the render methods to use html_writer, which
seems to make the code nicer.