Fix possible race condition in stack_cas_connection_db_cache.
The only way I could think of to fix this involves dropping the unique
index on hash. Because of teh way database transactions might overlap,
this seems to be unvoidable.

Now, we don't worry about inserting duplciate records, but we check for
duplicate records in get, and eliminate them when they are detected.