Instant validation of what the student types.
This uses ajax to submit what the student has typed, whenever they pause
in their typing.

There is an admin setting to turn this on/off, since it increases
server load, potentially quite a lot.

To try to reduce load, it rememebers every value it sends for validation
from the current page, so it does not need to sent the same things
twice.

At the moment, it does not work with Matrix elements. This can be fixed
later.