Fix maths output rendering.
There was a bug with how PRT feedback was displayed. We were calling
Moodle's format_text function before processing it as CAS text. That
does not work in situations where equations are being replaced by
images. (It does not matter when using MathJax.)

In order to fix this, there is now a restriction that all PRT node
feedback must have the same FORMAT_... type. In practice, this is not a
serious restriction, and the editing form now validates this.