Don't put the full plot URL in the CAS text from Maxima.
Instead we use a token !ploturl!, which is replace with the correct base
URL .../question/type/stack/plot.php/ in process_display_castext just
before the CAS text is output.

Note that I have left the URL_BASE variable in Maxima. This is safer for
people running Maxima on a separate server, it means that they don't
have to update the Maxima server at the same moment as the Moodle
server. At some point in the future, once everyone has upgraded, we can
remove URL_BASE.