Check the version of the Maxima library code used.
Since we have mechanisms like MaximaPool and maxima-optimised, it is
possible for the version of the STACK Maxima libraries being used to get
out of synch with the version that is shipped with the STACK question
type.

Here, we make every STACK evaluation report the version of the libraries
used. (It really needs to be every evaluation, since this is then stored
in the cache, and we would like to know when the cache gets out of
date.) We then check the version matches what is expected, and if not,
report it in two ways: 1. into the server's PHP logs, and 2. on the
healthcheck page.