Don't let users type $$ into the editing form.
Instead, there is a new option at the end of the form, to replace any
dollar delimiters when you save. This is useful if your are copying and
pasting from other sources of TeX, or just if you are used to typing $$
and $.

If that option is selected, it does what it says. If that option is not
selected, then the form validation will complain about any $$. However,
it will not complain about $, for the benefit of Americans who want to
ask about $2 + $2 = $4.