Front-end settings for how to display inverse trig functions.
Chris will need to link these up to the CAS code to make them work, but
there is a new field qtype_stack_options.inversetrig with three possible
values cos-1, acos or arccos.

This can be set on the editing form, and is correctly loaded and saved,
imported and exported, backed up and restored.