Fix issues with questions that have no PRTs and/or no inputs.
A STACK question with no inputs now behaves exactly like a Moodle
'Description' question.

A STACK question with inputs but no PRTs behaves like a Moodle Essay
question. If it is graded out of 0, then it has no effect on the score
at all. If it has a non-zero max mark, then it can be manually graded,
and the quiz total score is not computed until it has been.

Note that the manual graded case really ought to use a special
maualgradedwithexplicitvalidation behaviour, but life is too short, and
everyone should be using instant validation anyway.