Add an extra options field for input.
That is, add it in the database, on the form, and to import/export and
backup and restore.

This is necessary to make input types like dropdown-list work.

Note that this has not yet been wired up to to input type classes. There
are some new todos in the code for that.