Render a graphical representations of the PRTs in the editing form.
This uses SVG to render the graphs, which a laid out using an
algorithm I made up.

It does not work in Internet explorer, and so is hidden there.

Also, the links in the images do not work in webkit (Chrome & Safari)
but I don't know why. Even so, this is useful.

This change also improves the validation of the PRT structure.

The layout is at worse O(n^2) in the number of nodes. I actually think
that in common cases it is more likely to be O(n log n).

The unit tests are a bit sketchy. They are enough to verify that things
are basically working. For a better view of what is going on, for
example
to tune the layout algorithms, there are two test scripts:
* .../question/type/stack/stack/graphlayout/displayexamples.php
* .../question/type/stack/stack/graphlayout/displayprts.php