Copy quesion tests when doing 'Make copy' (Issue #41).
If you edit a question and then click the Make copy button
at the bottom of the form, we now detect this, and copy any question
tests from the original question to the new questions.

Of course, if you have added or removed a PRT or input, you will need to
fix the tests, but that is try if you edit a question normally too.