A script to run all the question tests for a context.
There is a new link on the STACK settings page. This takes you to a list
of all the contexts where there are STACK questions (a bit like the
index page for the replacedollars script.)

When you click on one of the links there, it runs all the question tests
for each deployed variant of each STACK qusetion there.

For undeployed questions, it runs the tests for one random variant.

The results give you links to go to the detailed results for that
variant of that question.