A script to tidy questions by renaming their parts.
This script lets you rename parts of a question, which is something you
cannot do using the normal question editing form because of the way the
form works. Also, once you see the work we have to do to rename things
throughout the question, you will understand why it needs to be a
separate option.

The script is accessed via a 'Tidy question' link next to the 'Question
tests & deployed versions' link.