Change docs so examples copy and paste better.
The key thing is that the maths in the example question text stays as
source, and is not rendered. We had been achieving that by using <pre>
tags, but they got copied into the Moodle HTML editor, which made the
question ugly. Here, I change the examples to be in a textarea, which
seems to work better.