take continued fraction approximation of the score at the end of a PRT
Inspired by my comment in maths/moodle-qtype_stack#183, this adds functions stack_utils::rational_approximation and stack_utils::fix_to_continued_fraction.
Instead of just rounding scores >= 0.9999999 up to 1, this finds a rational approximation to the score which agrees to 5 d.p. (need to decide the best number of d.p.)
This rounds numbers which are close to 1 up to 1, but also rounds e.g. 0.33333 to 1/3 (or as close as you can get to 1/3 in a float)

This needs unit tests, but I don't know how to write those.
It might make sense to replace stack_utils::fix_approximate_thirds with fix_to_continued_fraction - as far as I can see, that's only used on penalties at the moment.