Fix unit tests that use a provider #276
The data providers are run before pretty much anything else in the
PHPUnit run, and they are run whether or not that test will later be
filtered out. Therefore, it is critical not to do any serious processing
there.

Also, any processing using the CAS must be done after setUp is called,
which, which also means in the real test method.

Worse, since the CAS config is cached after the first call, this was
causing many other tests to fail too.