Always show question parts in the same order.
Before this change, on the Edit question form, inputs (and PRTs) were
shown in the order they appeared in the question text, while on the
Question tests page, they appeared in alphabetical order.

Not a terrible thing, but nicer to keep it consistent.

It seemed safest to get the order right when the question is
intitialised, even though this is a small amount of unnecessary work
in some cases. I don't think it is significant.