The JSXGraph block first iteration. Strings need to be extracted and proof reading done.
Requires validation for the input references and would be nice to gain access to the inputs to
do so.

Can't access the element identifiers of the inputs as those are only defined once the question
attempt exists and that does not exist when STACK renders the CASText so we now use a script
to find the inputs, not optimal but works.