Connect to YUI to tie the evaluation of JSXGraph code to the MathJax modules loading so that it can see the MathJax and does not try to use it before it exists.
Also revert to moving the code through the AMD function arguments even though it is against guides as if we do not take it out of the document the filtters will break it. Would be nice to move it through the dataset but as we cannot expect the filtters to be context aware and for example ignore attributes we cannot do that, hopefully implementations on other platforms can be more sensible.