Avoid time-outs and the consequences when clearing the CAS cache #373
While clearing the cache, we show a progressbar and call
core_php_time_limit::raise periodically. (This requires an extra
Continue click after the the cache has been purged using the button on
the Healthcheck page, but this actually makes it clearer that the cache
purge succeeded, so might be a good thing.

Clearing the cache now also deletes SVG plots.

During upgrade, we set_config('stackmaximaversion', ...) before clearing
the cache, because it is safter that way. (You can manually clear the
cache later, but you can't easily update that config value other than
directly in the database, if something goes wrong.