Hide the 'Question tests' link in iframes
Whe a STACK question is in an iframe, which might happen if you are
using filter_embedquestion, the it is a reall bad user-experience if you
can navigate to the 'Question tests & deployed versions' page.
Therefore, we hide the link in that case.

For embedded questions, authors will have to access the question tests
via the question bank.