Optimise various AST-filters, 10x speedups in some.
Current expected single AST-container setup time is 0.5ms for the parsing and 0.15ms for
filtering after that, currently ~10% of total time use is security and ~5% syntactical candy,
everything else just tries to add stars. As this is a lot it will make sense to cache the
question variables and other static content as preprocessed CAS-statements to skip the parsing
and validation, but for input processing we are fast enough. Expected gains from validation
caching in a reasonable sized question are in the order of 10-20ms once CASText2 comes into
play one can double that. With PRT-compilation we can however get an order of magnitude better
performance results.