Rename authoring_quick_start_quiz. Add information on question behaviours. Rename quick_start_4.
 Only suggest "next steps" at the very end.

- Rename authoring_quick_start_quiz to authoring_quick_start_8

This is a key guide that should be a part of the regular quick start guide

- Add information on the different question behaviours to the quiz guide

This is very helpful information, in my opinion

- Rename the guide on equivalence reasoning to Getting_started_with_equivalence_reasoning

This is a complex guide for a niche topic that is not required for an initial "quick start". Instead, I suggest it as a next step at the end of the guide.

- Minor edits to wording and structure
- Only suggest "next steps" at the very end of the guide series
Currently every guide ends with lots of examples of what to look at next. This makes it confusing and unclear which step is the recommended next step in the guide. The guide should now have a natural progression that ends in a lot of options.
git
- Made changes to progression and variable naming
- Prepared to add a screencast tutorial at the top of each quick start guide
This will help new users with simple problems such as finding the relevant sections and gives them two options for learning material; written or video.