Performance: reduce calls stack_maxima_latex_tidy > stack_string
This is a bit quick and dirty. The use of static here is dangerous. E.g.
this will break if one script tries to process things where some is in
one language and some is in another.

However, perf testing a quiz with 7 STACK questions, that caused
685 calls to stack_maxima_latex_tidy which caused ~7000 calls
to Moodle's get_string, which too about 1.5s of the page load time.
So, in the common case where everything is in one language, this is
a worthwhile performance win. If it causes bugs for some people, we
can fix that later.