JSXGraph 1.2.1, not tested.
Note for Alfred: the minified coming from normal build process is good, we just changed the name of the defined thing as we need to refrence it through the Moodle JavaScript logic. MAybe someone can arrange an alias to be defined so that things will work in a simpler way.
Note to Tim: Pushing the src-version of JSXGraph through the Moodle grunt process just does not work, too many code-style differences.