Mixed format (CASText2) document support. Put a MD-table inside HTML if you wish and the math mode detection and automatic wrapping and escaping of injections should be able to deal with it.
You will still need to tell us if you want to do it. Basically, wrap the content in `[[markdownformat]]..[[/markdownformat]]` if you wish to do so.

No matter what mode you put around `[[jsxgraph]]` the content will be handled as HTML-format and injections will still not directly work in JS, you still need to do escaping when mixing stuff there. This is simpler and we do not want to break backwards compatibility.