Update JSXGraph.md
Added a detailed example on region shading as a way to explore custom binding functions. I acknowledge that this may work better as a case study or as its own page in the docs - it is rather long - but I figured I would submit it here first. If you think it SHOULD be moved, I would add some suggestions for how to introduce randomisation to this question and discuss the difficulties with this type of question when the regions aren;t uniform like this.