Update security-map.json
Can we allow the variable "stringdisp" for teachers or are there any security or design concerns about this?
I never use string functions or variables for calculations, but sometimes they could be useful for feedback. I often define a ternary function like "ternary(bool, exp1, exp2)" anyway and it is sometimes clearer to use this function than an [[if]] [[else]] [[/if]] construction, e.g. in a singular/plural situation.
Example: "{#n#} parameters are correct" is grammatically wrong when n=1. So I like to write "{#n#} {#ternary(n=1,"parameter is", "parameters are")#} correct", but then I get unwanted quotes in the feedback. There are surely many ways to solve such problems, but an easy way would be to allow "stringdisp".